+++ src/python/re/index.ts
@@ -1,3 +1,4 @@
 export { findall } from './findall.ts'
 export { finditer } from './finditer.ts'
 export { sub } from './sub.ts'
+export { subn } from './subn.ts'